Add experimental FluentDb2 pt. 4.

+++ backend/sivujetti/src/Page/PagesRepository2.php
@@ -4,7 +4,7 @@ namespace Sivujetti\Page;
 
 use Envms\FluentPDO\Queries\Select;
 use Pike\{ArrayUtils, PikeException};
-use Pike\Db\{FluentDb, MyDelete, MyInsert, MyUpdate};
+use Pike\Db\{FluentDb, FluentDb2, MyDelete, MyInsert, MyUpdate, Query};
 use Pike\Interfaces\RowMapperInterface;
 use Sivujetti\Block\Entities\Block;
 use Sivujetti\Db\TempJsonCompatSelect;
@@ -17,23 +17,28 @@ final class PagesRepository2 {
     public const HARD_LIMIT = 200;
     /** @var \Pike\Db\FluentDb */
     private FluentDb $fluentDb;
+    /** @var \Pike\Db\FluentDb2 */
+    private FluentDb2 $fluentDb2;
     /** @var \ArrayObject<int, \Sivujetti\PageType\Entities\PageType> */
     private \ArrayObject $pageTypes;
     /**
      * @param \Pike\Db\FluentDb $fluentDb
+     * @param \Pike\Db\FluentDb2 $fluentDb2
      * @param \Sivujetti\TheWebsite\Entities\TheWebsite $theWebsite
      */
-    public function __construct(FluentDb $fluentDb, TheWebsite $theWebsite) {
+    public function __construct(FluentDb $fluentDb, FluentDb2 $fluentDb2, TheWebsite $theWebsite) {
         $this->fluentDb = $fluentDb;
+        $this->fluentDb2 = $fluentDb2;
         $this->pageTypes = $theWebsite->pageTypes;
     }
     /**
      * @param string $pageTypeName = "Pages"
      * @param string[] $fields = [] array<int, "@simple"|"@blocks">
-     * @return \Pike\Db\MySelect
+     * @return \Pike\Db\MySelect|\Pike\Db\Query
      */
-    public function select(string $pageTypeName = "Pages", array $fields = []): Select {
+    public function select(string $pageTypeName = "Pages", array $fields = []): Select|Query {
         $pageType = $this->getPageTypeOrThrow($pageTypeName);
+        if (!defined("USE_NEW_FLUENT_DB")) {
         $sqliteVersion = "3.37";
         $compatCls = version_compare($sqliteVersion, "3.38", ">=") ? null : TempJsonCompatSelect::class;
         return $this->fluentDb->select("\${p}{$pageType->name} p", Page::class, $compatCls)
@@ -50,29 +55,67 @@ final class PagesRepository2 {
                     return $page;
                 }
             });
+        } else {
+        return $this->fluentDb2->select("\${p}{$pageType->name} p")
+            ->fields(self::createSelectFields($fields, $pageType))
+            ->fetchWith(\PDO::FETCH_CLASS, Page::class)
+            ->mapWith(function (Page $row, int $_rowNum, array $_rows): ?Page {
+                // $row->slug set by pdo
+                // $row->path set by pdo
+                // $row->level set by pdo
+                // $row->title set by pdo
+                $row->meta = $row->metaJson ? JsonUtils::parse($row->metaJson) : null;
+                unset($row->metaJson);
+                // $row->layoutId set by pdo
+                // $row->id set by pdo
+                // $row->type set by pdo
+                $row->blocks = array_map(fn($blockRaw) =>
+                    Block::fromObject($blockRaw)
+                , $row->blocksJson ? JsonUtils::parse($row->blocksJson) : []);
+                unset($row->blocksJson);
+                // $row->status set by pdo
+                // $row->createdAt leave unset
+                // $row->lastUpdatedAt leave unset
+                // $row->layout leave unset
+                // /* Own props set by pdo */
+                return $row;
+            });
+        }
     }
     /**
      * @param string $pageTypeName = "Pages"
-     * @return \Pike\Db\MyInsert
+     * @return \Pike\Db\MyInsert|\Pike\Db\Query
      */
-    public function insert(PageType $pageType): MyInsert {
+    public function insert(PageType $pageType): MyInsert|Query {
+        if (!defined("USE_NEW_FLUENT_DB")) {
         return $this->fluentDb->insert("\${p}{$pageType->name}");
+        } else {
+        return $this->fluentDb2->insert("\${p}{$pageType->name}");
+        }
     }
     /**
      * @param string $pageTypeName = "Pages"
-     * @return \Pike\Db\MyUpdate
+     * @return \Pike\Db\MyUpdate|\Pike\Db\Query
      */
-    public function update(string $pageTypeName = "Pages"): MyUpdate {
+    public function update(string $pageTypeName = "Pages"): MyUpdate|Query {
         $pageType = $this->getPageTypeOrThrow($pageTypeName);
+        if (!defined("USE_NEW_FLUENT_DB")) {
         return $this->fluentDb->update("\${p}{$pageType->name}");
+        } else {
+        return $this->fluentDb2->update("\${p}{$pageType->name}");
+        }
     }
     /**
      * @param string $pageTypeName = "Pages"
-     * @return \Pike\Db\MyDelete
+     * @return \Pike\Db\MyDelete|\Pike\Db\Query
      */
-    public function delete(string $pageTypeName = "Pages"): MyDelete {
+    public function delete(string $pageTypeName = "Pages"): MyDelete|Query {
         $pageType = $this->getPageTypeOrThrow($pageTypeName);
+        if (!defined("USE_NEW_FLUENT_DB")) {
         return $this->fluentDb->delete("\${p}{$pageType->name}");
+        } else {
+        return $this->fluentDb2->delete("\${p}{$pageType->name}");
+        }
     }
     /**
      * @param string $candidate
@@ -98,7 +141,8 @@ final class PagesRepository2 {
             "title",
             "meta as metaJson",
             "layoutId",
-            "'{$pageType->name}' AS `type`", "status",
+            "'{$pageType->name}' AS `type`",
+            "status",
             (in_array("@blocks", $fields, true) ? "blocks" : "NULL") . " AS blocksJson",
         ];
         if (in_array("@own", $fields, true))
